WIP moving building of contracts to runtime instead of codegen
